Refactored controller directory
