+++ Code/Bachelor.Thesis.Benchmarking.WebApi/Bachelor.Thesis.Benchmarking.WebApi.csproj
@@ -20,8 +20,4 @@
 	<ItemGroup>
 		<ProjectReference Include="..\Bachelor.Thesis.Benchmarking\Bachelor.Thesis.Benchmarking.csproj" />
 	</ItemGroup>
-
-	<ItemGroup>
-	  <Folder Include="Controller\" />
-	</ItemGroup>
 </Project>
